+++ app/src/main/java/com/orgzly/android/link/OrgRoamIds.kt
@@ -1,0 +1,7 @@
+package com.orgzly.android.link
+
+import java.util.UUID
+
+object OrgRoamIds {
+    fun newId(): String = UUID.randomUUID().toString()
+}